+++ packages/forma-36-react-components/src/components/Illustration/Illustration.tsx
@@ -2,17 +2,17 @@ import React, { Component } from 'react';
 import cn from 'classnames';
 import { illustrationName } from './constants';
 
-const Archive = require('./svg/Archive.svg');
-const Audio = require('./svg/Audio.svg');
-const Code = require('./svg/Code.svg');
-const Image = require('./svg/Image.svg');
-const Markup = require('./svg/Markup.svg');
-const Pdf = require('./svg/Pdf.svg');
-const Plaintext = require('./svg/Plaintext.svg');
-const Presentation = require('./svg/Presentation.svg');
-const Richtext = require('./svg/Richtext.svg');
-const Spreadsheet = require('./svg/Spreadsheet.svg');
-const Video = require('./svg/Video.svg');
+import Archive from './svg/Archive.svg';
+import Audio from './svg/Audio.svg';
+import Code from './svg/Code.svg';
+import Image from './svg/Image.svg';
+import Markup from './svg/Markup.svg';
+import Pdf from './svg/Pdf.svg';
+import Plaintext from './svg/Plaintext.svg';
+import Presentation from './svg/Presentation.svg';
+import Richtext from './svg/Richtext.svg';
+import Spreadsheet from './svg/Spreadsheet.svg';
+import Video from './svg/Video.svg';
 
 import styles from './Illustration.css';