+++ packages/forma-36-react-components/src/components/Tag/Tag.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import cn from 'classnames';
 
-const styles = require('./Tag.css');
+import styles from './Tag.css';
 
 export type TagType =
   | 'primary'